@@ -1,5 +1,5 @@
 ﻿
-#pragma warning disable 649 // We don't care about unsused fields here, because they are mapped with the input file.
+#pragma warning disable 649 // We don't care about unused fields here, because they are mapped with the input file.
 
 using Microsoft.ML.Data;
 
@@ -9,5 +9,7 @@ namespace GitHubLabeler.DataStructures
     {
         [ColumnName("PredictedLabel")]
         public string Area;
+
+        public float[] Score;
     }
 }
