@@ -1,5 +1,5 @@
 ﻿
-#pragma warning disable 649 // We don't care about unsused fields here, because they are mapped with the input file.
+#pragma warning disable 649 // We don't care about unused fields here, because they are mapped with the input file.
 
 namespace GitHubLabeler.DataStructures
 {
