@@ -4,6 +4,7 @@
     {
         public int MovieID { get; set; }
         public string MovieName { get; set; }
-        public bool liked; 
-    }    
+
+        public bool liked;
+    }
 }
