@@ -1,13 +1,10 @@
-﻿using Microsoft.ML.Runtime.Api;
+﻿using Microsoft.ML.Data;
 
 namespace SpamDetectionConsoleApp.MLDataStructures
 {
     class SpamPrediction
     {
         [ColumnName("PredictedLabel")]
-        public bool isSpam { get; set; }
-
-        public float Score { get; set; }
-        public float Probability { get; set; }
+        public string isSpam { get; set; }
     }
 }
