@@ -1,5 +1,4 @@
-﻿using Microsoft.ML.Core.Data;
-using OxyPlot;
+﻿using OxyPlot;
 using OxyPlot.Series;
 using System;
 using System.Collections.Generic;
@@ -31,32 +30,26 @@ namespace CustomerSegmentation.Model
             _mlContext = mlContext;
         }
 
-        public ITransformer LoadModelFromZipFile(string modelPath)
+        public ITransformer LoadModel(string modelPath)
         {
-            using (var stream = new FileStream(modelPath, FileMode.Open, FileAccess.Read, FileShare.Read))
-            {
-                _trainedModel = _mlContext.Model.Load(stream);
-            }
-
+            _trainedModel = _mlContext.Model.Load(modelPath, out var modelInputSchema);
             return _trainedModel;
         }
 
         public void CreateCustomerClusters()
         {
-            TextLoader reader = _mlContext.Data.CreateTextReader(
+            var data = _mlContext.Data.LoadFromTextFile(path:_pivotDataLocation,
                             columns: new[]
                                         {
-                                          new TextLoader.Column("Features", DataKind.R4, new[] {new TextLoader.Range(0, 31) }),
-                                          new TextLoader.Column("LastName", DataKind.Text, 32)
+                                          new TextLoader.Column("Features", DataKind.Single, new[] {new TextLoader.Range(0, 31) }),
+                                          new TextLoader.Column(nameof(PivotData.LastName), DataKind.String, 32)
                                         },
                             hasHeader: true,
                             separatorChar: ',');
-
-            var data = reader.Read(_pivotDataLocation);
-
+            
             //Apply data transformation to create predictions/clustering
-            var predictions = _trainedModel.Transform(data)
-                            .AsEnumerable<ClusteringPrediction>(_mlContext, false)
+            var tranfomedDataView = _trainedModel.Transform(data);
+            var predictions = _mlContext.Data.CreateEnumerable <ClusteringPrediction>(tranfomedDataView, false)
                             .ToArray();
 
             //Generate data files with customer data grouped by clusters
