@@ -1,15 +1,10 @@
 ﻿using System;
 using System.IO;
-
 using Microsoft.ML;
-using Microsoft.ML.Core.Data;
-using Microsoft.ML.Transforms;
-using Microsoft.ML.Transforms.Categorical;
-using Microsoft.ML.Transforms.Projections;
-
 using CustomerSegmentation.DataStructures;
 using Common;
 using Microsoft.ML.Data;
+using Microsoft.ML.Transforms;
 
 namespace CustomerSegmentation
 {
@@ -17,12 +12,13 @@ namespace CustomerSegmentation
     {
         static void Main(string[] args)
         {
-            var assetsPath = PathHelper.GetAssetsPath(@"..\..\..\assets");
+            string assetsRelativePath = @"../../../assets";
+            string assetsPath = GetAbsolutePath(assetsRelativePath);
 
-            var transactionsCsv = Path.Combine(assetsPath, "inputs", "transactions.csv");
-            var offersCsv = Path.Combine(assetsPath, "inputs", "offers.csv");
-            var pivotCsv = Path.Combine(assetsPath, "inputs", "pivot.csv");
-            var modelZip = Path.Combine(assetsPath, "outputs", "retailClustering.zip");
+            string transactionsCsv = Path.Combine(assetsPath, "inputs", "transactions.csv");
+            string offersCsv = Path.Combine(assetsPath, "inputs", "offers.csv");
+            string pivotCsv = Path.Combine(assetsPath, "inputs", "pivot.csv");
+            string modelPath = Path.Combine(assetsPath, "outputs", "retailClustering.zip");
 
             try
             {
@@ -33,28 +29,26 @@ namespace CustomerSegmentation
                 MLContext mlContext = new MLContext(seed: 1);  //Seed set to any number so you have a deterministic environment
 
                 // STEP 1: Common data loading configuration
-                TextLoader textLoader = mlContext.Data.CreateTextReader(
-                                            columns:new[]
+                var pivotDataView = mlContext.Data.LoadFromTextFile(path: pivotCsv,
+                                            columns: new[]
                                                         {
-                                                        new TextLoader.Column("Features", DataKind.R4, new[] {new TextLoader.Range(0, 31) }),
-                                                        new TextLoader.Column("LastName", DataKind.Text, 32)
+                                                        new TextLoader.Column("Features", DataKind.Single, new[] {new TextLoader.Range(0, 31) }),
+                                                        new TextLoader.Column(nameof(PivotData.LastName), DataKind.String, 32)
                                                         },
                                             hasHeader: true,
                                             separatorChar: ',');
 
-                var pivotDataView = textLoader.Read(pivotCsv);
-
                 //STEP 2: Configure data transformations in pipeline
-                var dataProcessPipeline =  new PrincipalComponentAnalysisEstimator(mlContext, "Features", "PCAFeatures", rank: 2)
-                                                .Append(new OneHotEncodingEstimator(mlContext, new[] { new OneHotEncodingEstimator.ColumnInfo("LastName",
-                                                                                                                                              "LastNameKey",
-                                                                                                                                              OneHotEncodingTransformer.OutputKind.Ind) }));
-                // (Optional) Peek data in training DataView after applying the ProcessPipeline's transformations  
-                Common.ConsoleHelper.PeekDataViewInConsole<PivotObservation>(mlContext, pivotDataView, dataProcessPipeline, 10);
+                var dataProcessPipeline = mlContext.Transforms.ProjectToPrincipalComponents(outputColumnName: "PCAFeatures", inputColumnName: "Features", rank: 2)
+                 .Append(mlContext.Transforms.Categorical.OneHotEncoding(outputColumnName: "LastNameKey", inputColumnName: nameof(PivotData.LastName), OneHotEncodingEstimator.OutputKind.Indicator));
+
+
+                // (Optional) Peek data in training DataView after applying the ProcessPipeline's transformations
+                Common.ConsoleHelper.PeekDataViewInConsole(mlContext, pivotDataView, dataProcessPipeline, 10);
                 Common.ConsoleHelper.PeekVectorColumnDataInConsole(mlContext, "Features", pivotDataView, dataProcessPipeline, 10);
 
-                //STEP 3: Create the training pipeline                
-                var trainer = mlContext.Clustering.Trainers.KMeans("Features", clustersCount: 3);
+                //STEP 3: Create the training pipeline
+                var trainer = mlContext.Clustering.Trainers.KMeans(featureColumnName: "Features", numberOfClusters: 3);
                 var trainingPipeline = dataProcessPipeline.Append(trainer);
 
                 //STEP 4: Train the model fitting to the pivotDataView
@@ -64,22 +58,31 @@ namespace CustomerSegmentation
                 //STEP 5: Evaluate the model and show accuracy stats
                 Console.WriteLine("===== Evaluating Model's accuracy with Test data =====");
                 var predictions = trainedModel.Transform(pivotDataView);
-                var metrics = mlContext.Clustering.Evaluate(predictions, score: "Score", features: "Features");
+                var metrics = mlContext.Clustering.Evaluate(predictions, scoreColumnName: "Score", featureColumnName: "Features");
 
                 ConsoleHelper.PrintClusteringMetrics(trainer.ToString(), metrics);
 
                 //STEP 6: Save/persist the trained model to a .ZIP file
-                using (var fs = new FileStream(modelZip, FileMode.Create, FileAccess.Write, FileShare.Write))
-                    mlContext.Model.Save(trainedModel, fs);
+                mlContext.Model.Save(trainedModel, pivotDataView.Schema, modelPath);
 
-                Console.WriteLine("The model is saved to {0}", modelZip);
+                Console.WriteLine("The model is saved to {0}", modelPath);
             }
             catch (Exception ex)
             {
-                Common.ConsoleHelper.ConsoleWriteException(ex.Message);
+                Common.ConsoleHelper.ConsoleWriteException(ex.ToString());
             }
 
             Common.ConsoleHelper.ConsolePressAnyKey();
+
+        }
+        public static string GetAbsolutePath(string relativePath)
+        {
+            FileInfo _dataRoot = new FileInfo(typeof(Program).Assembly.Location);
+            string assemblyFolderPath = _dataRoot.Directory.FullName;
+
+            string fullPath = Path.Combine(assemblyFolderPath, relativePath);
+
+            return fullPath;
         }
     }
 }
