@@ -1,10 +1,8 @@
 ﻿using System;
 using System.IO;
-
 using Microsoft.ML;
 using Common;
 using Clustering_Iris.DataStructures;
-using Microsoft.ML.Core.Data;
 using Microsoft.ML.Data;
 
 namespace Clustering_Iris
@@ -13,56 +11,61 @@ namespace Clustering_Iris
     {
         private static string AppPath => Path.GetDirectoryName(Environment.GetCommandLineArgs()[0]);
 
-        private static string BaseDatasetsLocation = @"../../../../Data";
-        private static string DataPath = $"{BaseDatasetsLocation}/iris-full.txt";
+        private static string BaseDatasetsRelativePath = @"../../../../Data";
+        private static string DataSetRealtivePath = $"{BaseDatasetsRelativePath}/iris-full.txt";
 
-        private static string BaseModelsPath = @"../../../../MLModels";
-        private static string ModelPath = $"{BaseModelsPath}/IrisModel.zip";
+        private static string DataPath = GetAbsolutePath(DataSetRealtivePath);
+
+        private static string BaseModelsRelativePath = @"../../../../MLModels";
+        private static string ModelRelativePath = $"{BaseModelsRelativePath}/IrisModel.zip";
+
+        private static string ModelPath = GetAbsolutePath(ModelRelativePath);
+        private static IDataView trainingDataView;
+        private static IDataView testingDataView;
 
         private static void Main(string[] args)
         {
             //Create the MLContext to share across components for deterministic results
             MLContext mlContext = new MLContext(seed: 1);  //Seed set to any number so you have a deterministic environment
 
-            // STEP 1: Common data loading configuration
-            TextLoader textLoader = mlContext.Data.CreateTextReader(
+            // STEP 1: Common data loading configuration            
+            IDataView fullData = mlContext.Data.LoadFromTextFile(path: DataPath,
                                                 columns:new[]
                                                             {
-                                                                new TextLoader.Column("Label", DataKind.R4, 0),
-                                                                new TextLoader.Column("SepalLength", DataKind.R4, 1),
-                                                                new TextLoader.Column("SepalWidth", DataKind.R4, 2),
-                                                                new TextLoader.Column("PetalLength", DataKind.R4, 3),
-                                                                new TextLoader.Column("PetalWidth", DataKind.R4, 4),
+                                                                new TextLoader.Column("Label", DataKind.Single, 0),
+                                                                new TextLoader.Column(nameof(IrisData.SepalLength), DataKind.Single, 1),
+                                                                new TextLoader.Column(nameof(IrisData.SepalWidth), DataKind.Single, 2),
+                                                                new TextLoader.Column(nameof(IrisData.PetalLength), DataKind.Single, 3),
+                                                                new TextLoader.Column(nameof(IrisData.PetalWidth), DataKind.Single, 4),
                                                             },
                                                 hasHeader:true,
                                                 separatorChar:'\t');
 
-            IDataView fullData = textLoader.Read(DataPath);
-
             //Split dataset in two parts: TrainingDataset (80%) and TestDataset (20%)
-            (IDataView trainingDataView, IDataView testingDataView) = mlContext.Clustering.TrainTestSplit(fullData, testFraction: 0.2);
+            DataOperationsCatalog.TrainTestData trainTestData = mlContext.Data.TrainTestSplit(fullData, testFraction: 0.2);
+            trainingDataView = trainTestData.TrainSet;
+            testingDataView = trainTestData.TestSet;
 
             //STEP 2: Process data transformations in pipeline
-            var dataProcessPipeline = mlContext.Transforms.Concatenate("Features", "SepalLength", "SepalWidth", "PetalLength", "PetalWidth");
+            var dataProcessPipeline = mlContext.Transforms.Concatenate("Features", nameof(IrisData.SepalLength), nameof(IrisData.SepalWidth), nameof(IrisData.PetalLength), nameof(IrisData.PetalWidth));
 
             // (Optional) Peek data in training DataView after applying the ProcessPipeline's transformations  
-            Common.ConsoleHelper.PeekDataViewInConsole<IrisData>(mlContext, trainingDataView, dataProcessPipeline, 10);
+            Common.ConsoleHelper.PeekDataViewInConsole(mlContext, trainingDataView, dataProcessPipeline, 10);
             Common.ConsoleHelper.PeekVectorColumnDataInConsole(mlContext, "Features", trainingDataView, dataProcessPipeline, 10);
 
             // STEP 3: Create and train the model     
-            var trainer = mlContext.Clustering.Trainers.KMeans(features: "Features", clustersCount: 3);
+            var trainer = mlContext.Clustering.Trainers.KMeans(featureColumnName: "Features", numberOfClusters: 3);
             var trainingPipeline = dataProcessPipeline.Append(trainer);
             var trainedModel = trainingPipeline.Fit(trainingDataView);
 
             // STEP4: Evaluate accuracy of the model
             IDataView predictions = trainedModel.Transform(testingDataView);
-            var metrics = mlContext.Clustering.Evaluate(predictions, score: "Score", features: "Features");
+            var metrics = mlContext.Clustering.Evaluate(predictions, scoreColumnName: "Score", featureColumnName: "Features");
 
             ConsoleHelper.PrintClusteringMetrics(trainer.ToString(), metrics);
 
             // STEP5: Save/persist the model as a .ZIP file
-            using (var fs = new FileStream(ModelPath, FileMode.Create, FileAccess.Write, FileShare.Write))
-                mlContext.Model.Save(trainedModel, fs);
+            mlContext.Model.Save(trainedModel, trainingDataView.Schema, ModelPath);
 
             Console.WriteLine("=============== End of training process ===============");
 
@@ -77,20 +80,27 @@ namespace Clustering_Iris
                 PetalWidth = 5.1f,
             };
             
-            using (var stream = new FileStream(ModelPath, FileMode.Open, FileAccess.Read, FileShare.Read))
-            {
-                ITransformer model = mlContext.Model.Load(stream);
-                // Create prediction engine related to the loaded trained model
-                var predEngine = model.CreatePredictionEngine<IrisData, IrisPrediction>(mlContext);
+            ITransformer model = mlContext.Model.Load(ModelPath, out var modelInputSchema);
+            // Create prediction engine related to the loaded trained model
+            var predEngine = mlContext.Model.CreatePredictionEngine<IrisData, IrisPrediction>(model);
 
-                //Score
-                var resultprediction = predEngine.Predict(sampleIrisData);
+            //Score
+            var resultprediction = predEngine.Predict(sampleIrisData);
 
-                Console.WriteLine($"Cluster assigned for setosa flowers:" + resultprediction.SelectedClusterId);
-            }
-
+            Console.WriteLine($"Cluster assigned for setosa flowers:" + resultprediction.SelectedClusterId);
+          
             Console.WriteLine("=============== End of process, hit any key to finish ===============");
             Console.ReadKey();           
+        }
+
+        public static string GetAbsolutePath(string relativePath)
+        {
+            FileInfo _dataRoot = new FileInfo(typeof(Program).Assembly.Location);
+            string assemblyFolderPath = _dataRoot.Directory.FullName;
+
+            string fullPath = Path.Combine(assemblyFolderPath, relativePath);
+
+            return fullPath;
         }
     }
 
