@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Microsoft.ML.Data;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -7,8 +8,10 @@ namespace ImageClassification.ImageDataStructures
 {
     public class ImageNetData
     {
+        [LoadColumn(0)]
         public string ImagePath;
 
+        [LoadColumn(1)]
         public string Label;
 
         public static IEnumerable<ImageNetData> ReadFromCsv(string file, string folder)
