@@ -9,15 +9,13 @@ namespace ImageClassification
     {
         static void Main(string[] args)
         {
-            var assetsPath = ModelHelpers.GetAssetsPath(@"..\..\..\assets");
+            string assetsRelativePath = @"../../../assets";
+            string assetsPath = GetAbsolutePath(assetsRelativePath);
 
             var tagsTsv = Path.Combine(assetsPath, "inputs", "images", "tags.tsv");
             var imagesFolder = Path.Combine(assetsPath, "inputs", "images");
             var inceptionPb = Path.Combine(assetsPath, "inputs", "inception", "tensorflow_inception_graph.pb");
             var labelsTxt = Path.Combine(assetsPath, "inputs", "inception", "imagenet_comp_graph_label_strings.txt");
-
-            var customInceptionPb = Path.Combine(assetsPath, "inputs", "inception_custom", "model_tf.pb");
-            var customLabelsTxt = Path.Combine(assetsPath, "inputs", "inception_custom", "labels.txt");
 
             try
             {
@@ -27,10 +25,18 @@ namespace ImageClassification
             }
             catch (Exception ex)
             {
-                ConsoleHelpers.ConsoleWriteException(ex.Message);
+                ConsoleHelpers.ConsoleWriteException(ex.ToString());
             }
 
             ConsoleHelpers.ConsolePressAnyKey();
+        }
+
+        public static string GetAbsolutePath(string relativePath)
+        {
+            FileInfo _dataRoot = new FileInfo(typeof(Program).Assembly.Location);
+            string assemblyFolderPath = _dataRoot.Directory.FullName;
+            string fullPath = Path.Combine(assemblyFolderPath, relativePath);
+            return fullPath;
         }
     }
 }
