@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Microsoft.ML.Data;
+using System;
 using System.Collections.Generic;
 using System.Text;
 
@@ -6,10 +7,13 @@ namespace MovieRecommendationConsoleApp.DataStructures
 {
     public class MovieRating
     {
+        [LoadColumn(0)]
         public float userId;
 
+        [LoadColumn(1)]
         public float movieId;
 
+        [LoadColumn(2)]
         public float Label;
     }
 }
