@@ -7,31 +7,35 @@ namespace BikeSharingDemand.DataStructures
 {
     public class DemandObservation
     {
-        [LoadColumn(0)]
-        public float Season { get; set; }
-        [LoadColumn(1)]
-        public float Year { get; set; }
+        // Note that we're loading only some columns (certain indexes) starting on column number 2
+        // Also, the label column is number 16. 
+        // Columns 14, 15 are not being loaded from the file.
         [LoadColumn(2)]
-        public float Month { get; set; }
+        public float Season { get; set; }
         [LoadColumn(3)]
-        public float Hour { get; set; }
+        public float Year { get; set; }
         [LoadColumn(4)]
-        public float Holiday { get; set; }
+        public float Month { get; set; }
         [LoadColumn(5)]
-        public float Weekday { get; set; }
+        public float Hour { get; set; }
         [LoadColumn(6)]
-        public float WorkingDay { get; set; }
+        public float Holiday { get; set; }
         [LoadColumn(7)]
-        public float Weather { get; set; }
+        public float Weekday { get; set; }
         [LoadColumn(8)]
-        public float Temperature { get; set; }
+        public float WorkingDay { get; set; }
         [LoadColumn(9)]
-        public float NormalizedTemperature { get; set; }
+        public float Weather { get; set; }
         [LoadColumn(10)]
-        public float Humidity { get; set; }
+        public float Temperature { get; set; }
         [LoadColumn(11)]
-        public float Windspeed { get; set; }
+        public float NormalizedTemperature { get; set; }
         [LoadColumn(12)]
+        public float Humidity { get; set; }
+        [LoadColumn(13)]
+        public float Windspeed { get; set; }
+        [LoadColumn(16)]
+        [ColumnName("Label")]
         public float Count { get; set; }   // This is the observed count, to be used a "label" to predict
     }
 
