@@ -1,13 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
+using System.Globalization;
 using System.IO;
 using System.Linq;
 
-using Microsoft.ML;
-using Microsoft.ML.Core.Data;
-
 using BikeSharingDemand.DataStructures;
-using Common;
+
+using Microsoft.ML;
 
 namespace BikeSharingDemand
 {
@@ -43,19 +41,19 @@ namespace BikeSharingDemand
                 .Select(x => x.Split(','))
                 .Select(x => new DemandObservation()
                 {
-                    Season = float.Parse(x[2]),
-                    Year = float.Parse(x[3]),
-                    Month = float.Parse(x[4]),
-                    Hour = float.Parse(x[5]),
-                    Holiday = float.Parse(x[6]),
-                    Weekday = float.Parse(x[7]),
-                    WorkingDay = float.Parse(x[8]),
-                    Weather = float.Parse(x[9]),
-                    Temperature = float.Parse(x[10]),
-                    NormalizedTemperature = float.Parse(x[11]),
-                    Humidity = float.Parse(x[12]),
-                    Windspeed = float.Parse(x[13]),
-                    Count = float.Parse(x[16])
+                    Season = float.Parse(x[2], CultureInfo.InvariantCulture),
+                    Year = float.Parse(x[3], CultureInfo.InvariantCulture),
+                    Month = float.Parse(x[4], CultureInfo.InvariantCulture),
+                    Hour = float.Parse(x[5], CultureInfo.InvariantCulture),
+                    Holiday = float.Parse(x[6], CultureInfo.InvariantCulture),
+                    Weekday = float.Parse(x[7], CultureInfo.InvariantCulture),
+                    WorkingDay = float.Parse(x[8], CultureInfo.InvariantCulture),
+                    Weather = float.Parse(x[9], CultureInfo.InvariantCulture),
+                    Temperature = float.Parse(x[10], CultureInfo.InvariantCulture),
+                    NormalizedTemperature = float.Parse(x[11], CultureInfo.InvariantCulture),
+                    Humidity = float.Parse(x[12], CultureInfo.InvariantCulture),
+                    Windspeed = float.Parse(x[13], CultureInfo.InvariantCulture),
+                    Count = float.Parse(x[16], CultureInfo.InvariantCulture)
                 })
                 .Take(numberOfRecordsToRead)
                 .ToList();
