@@ -1,11 +1,9 @@
 ﻿using System;
 using Microsoft.ML;
-using Microsoft.ML.Core.Data;
 using System.IO;
-using Microsoft.ML.Data;
-
 using BikeSharingDemand.DataStructures;
 using Common;
+using Microsoft.ML.Data;
 
 namespace BikeSharingDemand
 {
@@ -14,8 +12,11 @@ namespace BikeSharingDemand
         private static string ModelsLocation = @"../../../../MLModels";
 
         private static string DatasetsLocation = @"../../../../Data";
-        private static string TrainingDataLocation = $"{DatasetsLocation}/hour_train.csv";
-        private static string TestDataLocation = $"{DatasetsLocation}/hour_test.csv";
+        private static string TrainingDataRelativePath = $"{DatasetsLocation}/hour_train.csv";
+        private static string TestDataRelativePath = $"{DatasetsLocation}/hour_test.csv";
+
+        private static string TrainingDataLocation = GetAbsolutePath(TrainingDataRelativePath);
+        private static string TestDataLocation = GetAbsolutePath(TestDataRelativePath);
         
         static void Main(string[] args)
         {
@@ -24,39 +25,23 @@ namespace BikeSharingDemand
             var mlContext = new MLContext(seed: 0);
 
             // 1. Common data loading configuration
-            var textLoader = mlContext.Data.CreateTextReader(
-                                                    columns:new[]
-                                                            {
-                                                            new TextLoader.Column("Season", DataKind.R4, 2),
-                                                            new TextLoader.Column("Year", DataKind.R4, 3),
-                                                            new TextLoader.Column("Month", DataKind.R4, 4),
-                                                            new TextLoader.Column("Hour", DataKind.R4, 5),
-                                                            new TextLoader.Column("Holiday", DataKind.R4, 6),
-                                                            new TextLoader.Column("Weekday", DataKind.R4, 7),
-                                                            new TextLoader.Column("WorkingDay", DataKind.R4, 8),
-                                                            new TextLoader.Column("Weather", DataKind.R4, 9),
-                                                            new TextLoader.Column("Temperature", DataKind.R4, 10),
-                                                            new TextLoader.Column("NormalizedTemperature", DataKind.R4, 11),
-                                                            new TextLoader.Column("Humidity", DataKind.R4, 12),
-                                                            new TextLoader.Column("Windspeed", DataKind.R4, 13),
-                                                            new TextLoader.Column("Count", DataKind.R4, 16)
-                                                            },
-                                                    hasHeader: true,
-                                                    separatorChar: ',');
-              
-            var trainingDataView = textLoader.Read(TrainingDataLocation);
-            var testDataView = textLoader.Read(TestDataLocation);
+            var trainingDataView = mlContext.Data.LoadFromTextFile<DemandObservation>(path: TrainingDataLocation, hasHeader:true, separatorChar: ',');
+            var testDataView = mlContext.Data.LoadFromTextFile<DemandObservation>(path: TestDataLocation, hasHeader:true, separatorChar: ',');
 
             // 2. Common data pre-process with pipeline data transformations
-            var dataProcessPipeline = mlContext.Transforms.CopyColumns("Count", "Label")
-                        // Concatenate all the numeric columns into a single features column
-                        .Append(mlContext.Transforms.Concatenate("Features", "Season", "Year", "Month",
-                                                                            "Hour", "Holiday", "Weekday",
-                                                                            "Weather", "Temperature", "NormalizedTemperature",
-                                                                            "Humidity", "Windspeed"));
+
+            // Concatenate all the numeric columns into a single features column
+            var dataProcessPipeline = mlContext.Transforms.Concatenate("Features",
+                                                     nameof(DemandObservation.Season), nameof(DemandObservation.Year), nameof(DemandObservation.Month),
+                                                     nameof(DemandObservation.Hour), nameof(DemandObservation.Holiday), nameof(DemandObservation.Weekday),
+                                                     nameof(DemandObservation.WorkingDay), nameof(DemandObservation.Weather), nameof(DemandObservation.Temperature),
+                                                     nameof(DemandObservation.NormalizedTemperature), nameof(DemandObservation.Humidity), nameof(DemandObservation.Windspeed))
+                                         .AppendCacheCheckpoint(mlContext);
+                                        // Use in-memory cache for small/medium datasets to lower training time. 
+                                        // Do NOT use it (remove .AppendCacheCheckpoint()) when handling very large datasets.
 
             // (Optional) Peek data in training DataView after applying the ProcessPipeline's transformations  
-            Common.ConsoleHelper.PeekDataViewInConsole<DemandObservation>(mlContext, trainingDataView, dataProcessPipeline, 10);
+            Common.ConsoleHelper.PeekDataViewInConsole(mlContext, trainingDataView, dataProcessPipeline, 10);
             Common.ConsoleHelper.PeekVectorColumnDataInConsole(mlContext, "Features", trainingDataView, dataProcessPipeline, 10);
 
             // Definition of regression trainers/algorithms to use
@@ -64,8 +49,8 @@ namespace BikeSharingDemand
             (string name, IEstimator<ITransformer> value)[] regressionLearners =
             {
                 ("FastTree", mlContext.Regression.Trainers.FastTree()),
-                ("Poisson", mlContext.Regression.Trainers.PoissonRegression()),
-                ("SDCA", mlContext.Regression.Trainers.StochasticDualCoordinateAscent()),
+                ("Poisson", mlContext.Regression.Trainers.LbfgsPoissonRegression()),
+                ("SDCA", mlContext.Regression.Trainers.Sdca()),
                 ("FastTreeTweedie", mlContext.Regression.Trainers.FastTreeTweedie()),
                 //Other possible learners that could be included
                 //...FastForestRegressor...
@@ -75,22 +60,21 @@ namespace BikeSharingDemand
 
             // 3. Phase for Training, Evaluation and model file persistence
             // Per each regression trainer: Train, Evaluate, and Save a different model
-            foreach (var learner in regressionLearners)
+            foreach (var trainer in regressionLearners)
             {
                 Console.WriteLine("=============== Training the current model ===============");
-                var trainingPipeline = dataProcessPipeline.Append(learner.value);
+                var trainingPipeline = dataProcessPipeline.Append(trainer.value);
                 var trainedModel = trainingPipeline.Fit(trainingDataView);
 
                 Console.WriteLine("===== Evaluating Model's accuracy with Test data =====");
                 IDataView predictions = trainedModel.Transform(testDataView);
-                var metrics = mlContext.Regression.Evaluate(predictions, label: "Count", score: "Score");               
-                ConsoleHelper.PrintRegressionMetrics(learner.value.ToString(), metrics);
+                var metrics = mlContext.Regression.Evaluate(data:predictions, labelColumnName:"Label", scoreColumnName: "Score");               
+                ConsoleHelper.PrintRegressionMetrics(trainer.value.ToString(), metrics);
 
                 //Save the model file that can be used by any application
-                string modelPath = $"{ModelsLocation}/{learner.name}Model.zip";
-                using (var fs = new FileStream(modelPath, FileMode.Create, FileAccess.Write, FileShare.Write))
-                    mlContext.Model.Save(trainedModel, fs);
-
+                string modelRelativeLocation = $"{ModelsLocation}/{trainer.name}Model.zip";
+                string modelPath = GetAbsolutePath(modelRelativeLocation);
+                mlContext.Model.Save(trainedModel, trainingDataView.Schema, modelPath);
                 Console.WriteLine("The model is saved to {0}", modelPath);
             }
 
@@ -101,15 +85,12 @@ namespace BikeSharingDemand
             foreach (var learner in regressionLearners)
             {
                 //Load current model from .ZIP file
-                ITransformer trainedModel;
-                string modelPath = $"{ModelsLocation}/{learner.name}Model.zip";
-                using (var stream = new FileStream(modelPath, FileMode.Open, FileAccess.Read, FileShare.Read))
-                {
-                    trainedModel = mlContext.Model.Load(stream);
-                }
+                string modelRelativeLocation = $"{ModelsLocation}/{learner.name}Model.zip";
+                string modelPath = GetAbsolutePath(modelRelativeLocation);
+                ITransformer trainedModel = mlContext.Model.Load(modelPath, out var modelInputSchema);
 
                 // Create prediction engine related to the loaded trained model
-                var predEngine = trainedModel.CreatePredictionEngine<DemandObservation, DemandPrediction>(mlContext);
+                var predEngine = mlContext.Model.CreatePredictionEngine<DemandObservation, DemandPrediction>(trainedModel);
 
                 Console.WriteLine($"================== Visualize/test 10 predictions for model {learner.name}Model.zip ==================");
                 //Visualize 10 tests comparing prediction with actual/observed values from the test dataset
@@ -117,7 +98,16 @@ namespace BikeSharingDemand
             }
 
             Common.ConsoleHelper.ConsolePressAnyKey();
+        }
 
+        public static string GetAbsolutePath(string relativePath)
+        {
+            FileInfo _dataRoot = new FileInfo(typeof(Program).Assembly.Location);
+            string assemblyFolderPath = _dataRoot.Directory.FullName;
+
+            string fullPath = Path.Combine(assemblyFolderPath, relativePath);
+
+            return fullPath;
         }
     }
 }
