@@ -1,20 +1,29 @@
 ﻿
+using Microsoft.ML.Data;
+
 namespace Regression_TaxiFarePrediction.DataStructures
 {
     public class TaxiTrip
     {
+        [LoadColumn(0)]
         public string VendorId;
 
+        [LoadColumn(1)]
         public string RateCode;
-        
+
+        [LoadColumn(2)]
         public float PassengerCount;
-        
+
+        [LoadColumn(3)]
         public float TripTime;
 
+        [LoadColumn(4)]
         public float TripDistance;
-        
+
+        [LoadColumn(5)]
         public string PaymentType;
 
+        [LoadColumn(6)]
         public float FareAmount;
     }
 }
