@@ -3,10 +3,8 @@ namespace Common
 module ConsoleHelper =
     open System
     open Microsoft.ML
-    open Microsoft.ML.Runtime.Data
     open Microsoft.ML.Data
-    open Microsoft.ML.Core.Data
-    open Microsoft.ML.Runtime.Api
+    //open Microsoft.ML.Api
     open System.Reflection
 
     let printPrediction prediction =
@@ -16,36 +14,44 @@ module ConsoleHelper =
 
     let printRegressionPredictionVersusObserved predictionCount observedCount =
         printfn "-------------------------------------------------"
-        printfn "Predicted : %d" predictionCount
+        printfn "Predicted : %.4f" predictionCount
         printfn "Actual:     %s" observedCount
         printfn "-------------------------------------------------"
 
-    let printRegressionMetrics name (metrics : RegressionEvaluator.Result) =
+    let printRegressionMetrics name (metrics : RegressionMetrics) =
         printfn "*************************************************"
         printfn "*       Metrics for %s regression model      " name
         printfn "*------------------------------------------------"
-        printfn "*       LossFn:        %.2f" metrics.LossFn
+        printfn "*       LossFn:        %.2f" metrics.LossFunction
         printfn "*       R2 Score:      %.2f" metrics.RSquared
-        printfn "*       Absolute loss: %.2f" metrics.L1
-        printfn "*       Squared loss:  %.2f" metrics.L2
-        printfn "*       RMS loss:      %.2f" metrics.Rms
+        printfn "*       Absolute loss: %.2f" metrics.MeanAbsoluteError
+        printfn "*       Squared loss:  %.2f" metrics.MeanSquaredError
+        printfn "*       RMS loss:      %.2f" metrics.RootMeanSquaredError
         printfn "*************************************************"
     
-    let printBinaryClassificationMetrics name (metrics : BinaryClassifierEvaluator.Result) =
+    let printBinaryClassificationMetrics name (metrics : CalibratedBinaryClassificationMetrics) =
         printfn"************************************************************"
         printfn"*       Metrics for %s binary classification model      " name
         printfn"*-----------------------------------------------------------"
         printfn"*       Accuracy: %.2f%%" (metrics.Accuracy * 100.)
-        printfn"*       Auc:      %.2f%%" (metrics.Auc * 100.)
+        printfn"*       Area Under Curve:      %.2f%%" (metrics.AreaUnderRocCurve * 100.)
+        printfn"*       Area under Precision recall Curve:    %.2f%%" (metrics.AreaUnderPrecisionRecallCurve * 100.)
         printfn"*       F1Score:  %.2f%%" (metrics.F1Score * 100.)
+ 
+        printfn"*       LogLogg:  %.2f%%" (metrics.LogLoss)
+        printfn"*       LogLossreduction:  %.2f%%" (metrics.LogLossReduction)
+        printfn"*       PositivePrecision:      %.2f" (metrics.PositivePrecision)
+        printfn"*       PositiveRecall:      %.2f" (metrics.PositiveRecall)
+        printfn"*       NegativePrecision:      %.2f" (metrics.NegativePrecision)
+        printfn"*       NegativeRecall:      %.2f" (metrics.NegativeRecall)
         printfn"************************************************************"
 
-    let printMultiClassClassificationMetrics name (metrics : MultiClassClassifierEvaluator.Result) =
+    let printMultiClassClassificationMetrics name (metrics : MulticlassClassificationMetrics) =
         printfn "************************************************************"
         printfn "*    Metrics for %s multi-class classification model   " name
         printfn "*-----------------------------------------------------------"
-        printfn "    AccuracyMacro = %.4f, a value between 0 and 1, the closer to 1, the better" metrics.AccuracyMacro
-        printfn "    AccuracyMicro = %.4f, a value between 0 and 1, the closer to 1, the better" metrics.AccuracyMicro
+        printfn "    AccuracyMacro = %.4f, a value between 0 and 1, the closer to 1, the better" metrics.MacroAccuracy
+        printfn "    AccuracyMicro = %.4f, a value between 0 and 1, the closer to 1, the better" metrics.MacroAccuracy
         printfn "    LogLoss = %.4f, the closer to 0, the better" metrics.LogLoss
         printfn "    LogLoss for class 1 = %.4f, the closer to 0, the better" metrics.PerClassLogLoss.[0]
         printfn "    LogLoss for class 2 = %.4f, the closer to 0, the better" metrics.PerClassLogLoss.[1]
@@ -63,16 +69,16 @@ module ConsoleHelper =
         let confidenceInterval95 = 1.96 * calculateStandardDeviation(values) / Math.Sqrt(float (values.Length-1));
         confidenceInterval95
 
-    let printMulticlassClassificationFoldsAverageMetrics algorithmName (crossValResults : (MultiClassClassifierEvaluator.Result * ITransformer * IDataView) array) =
+    let printMulticlassClassificationFoldsAverageMetrics algorithmName (crossValResults : TrainCatalogBase.CrossValidationResult<MulticlassClassificationMetrics>[]) =
         
-        let metricsInMultipleFolds = crossValResults |> Array.map(fun (metrics, model, scoredTestData) -> metrics)
+        let metricsInMultipleFolds = crossValResults |> Array.map(fun r -> r.Metrics)
 
-        let microAccuracyValues  = metricsInMultipleFolds |> Array.map(fun m -> m.AccuracyMicro)
+        let microAccuracyValues  = metricsInMultipleFolds |> Array.map(fun m -> m.MicroAccuracy)
         let microAccuracyAverage = microAccuracyValues |> Array.average
         let microAccuraciesStdDeviation = calculateStandardDeviation microAccuracyValues
         let microAccuraciesConfidenceInterval95 = calculateConfidenceInterval95 microAccuracyValues
 
-        let macroAccuracyValues = metricsInMultipleFolds |> Array.map(fun m -> m.AccuracyMacro)
+        let macroAccuracyValues = metricsInMultipleFolds |> Array.map(fun m -> m.MicroAccuracy)
         let macroAccuracyAverage = macroAccuracyValues |> Array.average
         let macroAccuraciesStdDeviation = calculateStandardDeviation macroAccuracyValues
         let macroAccuraciesConfidenceInterval95 = calculateConfidenceInterval95 macroAccuracyValues
@@ -96,12 +102,12 @@ module ConsoleHelper =
         printfn "*       Average LogLossReduction: %.3f  - Standard deviation: (%.3f)  - Confidence Interval 95%%: (%.3f)" logLossReductionAverage logLossReductionStdDeviation logLossReductionConfidenceInterval95
         printfn "*************************************************************************************************************"
 
-    let printClusteringMetrics name (metrics : ClusteringEvaluator.Result) =
+    let printClusteringMetrics name (metrics : ClusteringMetrics) =
         printfn "*************************************************"
         printfn "*       Metrics for %s clustering model      " name
         printfn "*------------------------------------------------"
-        printfn "*       AvgMinScore: %.4f" metrics.AvgMinScore
-        printfn "*       DBI is: %.4f" metrics.Dbi
+        printfn "*       Average Distance: %.15f" metrics.AverageDistance
+        printfn "*       Davies Bouldin Index is: %.15f" metrics.DaviesBouldinIndex
         printfn "*************************************************"
 
     let consoleWriteHeader line =
@@ -121,34 +127,25 @@ module ConsoleHelper =
 
     let peekDataViewInConsole<'TObservation when 'TObservation : (new : unit -> 'TObservation) and 'TObservation : not struct> (mlContext : MLContext) (dataView : IDataView) (pipeline : IEstimator<ITransformer>) numberOfRows =
         
-        let msg = sprintf "Peek data in DataView: Showing %d rows with the columns specified by TObservation class" numberOfRows
+        let msg = sprintf "Peek data in DataView: Showing %d rows with the columns" numberOfRows
         consoleWriteHeader msg
 
         //https://github.com/dotnet/machinelearning/blob/master/docs/code/MlNetCookBook.md#how-do-i-look-at-the-intermediate-data
         let transformer = pipeline.Fit dataView
         let transformedData = transformer.Transform dataView
 
-        // 'transformedData' is a 'promise' of data, lazy-loading. Let's actually read it.
-        // Convert to an enumerable of user-defined type.
-        let someRows = 
-            transformedData.AsEnumerable<'TObservation>(mlContext, reuseRowObject = false)
-            // Take the specified number of rows
-            |> Seq.take numberOfRows
-            // Convert to List
-            |> Seq.toList
+        // 'transformedData' is a 'promise' of data, lazy-loading. call Preview  
+        //and iterate through the returned collection from preview.
 
-        someRows
-        |> List.iter(fun row ->
-                        
-                        let lineToPrint =
-                            row.GetType().GetFields(BindingFlags.Instance ||| BindingFlags.Static ||| BindingFlags.NonPublic ||| BindingFlags.Public)
-                            |> Array.map(fun field -> sprintf "| %s: %O" field.Name (field.GetValue(row)))
-                            |> Array.fold (+) "Row--> "
-
-                        printfn "%s" lineToPrint
+        transformedData.Preview(numberOfRows).RowView
+        |> Seq.iter 
+            (fun row ->
+                row.Values
+                |> Array.map (function KeyValue(k,v) -> sprintf "| %s:%O" k v)
+                |> Array.fold (+) "Row--> "
+                |> printfn "%s\n"
             )
-    
-        someRows
+
 
     let peekVectorColumnDataInConsole (mlContext : MLContext) columnName (dataView : IDataView) (pipeline : IEstimator<ITransformer>) numberOfRows =
         let msg = sprintf "Peek data in DataView: : Show %d rows with just the '%s' column" numberOfRows columnName
@@ -159,7 +156,7 @@ module ConsoleHelper =
 
         // Extract the 'Features' column.
         let someColumnData = 
-            transformedData.GetColumn<float32[]>(mlContext, columnName)
+            transformedData.GetColumn<float32[]>(columnName)
             |> Seq.take numberOfRows
             |> Seq.toList
 
@@ -169,7 +166,7 @@ module ConsoleHelper =
             let concatColumn = 
                 row
                 |> Array.map string
-                |> Array.fold (+) " "
+                |> String.concat ""
             printfn "%s" concatColumn
         )
                         
