@@ -1,12 +1,16 @@
 ﻿namespace SentimentAnalysis.DataStructures
 
 module Model =
-    open Microsoft.ML.Runtime.Api
+    open Microsoft.ML.Data
 
     /// Type representing the text to run sentiment analysis on.
     [<CLIMutable>] 
     type SentimentIssue = 
         { 
+            [<LoadColumn(0)>]
+            Label : bool
+
+            [<LoadColumn(2)>]
             Text : string 
         }
 
@@ -14,9 +18,14 @@ module Model =
     [<CLIMutable>]
     type  SentimentPrediction = 
         { 
-            // Predicted sentiment: 0 - negative, 1 - positive
+            // ColumnName attribute is used to change the column name from
+            // its default value, which is the name of the field.
             [<ColumnName("PredictedLabel")>]
             Prediction : bool; 
+
+            // No need to specify ColumnName attribute, because the field
+            // name "Probability" is the column name we want.
             Probability : float32; 
+
             Score : float32 
         }
