@@ -3,7 +3,7 @@
 open System
 open System.IO
 open Microsoft.ML
-open Microsoft.ML.Runtime.Data
+open Microsoft.ML.Data
 open Clustering_Iris.DataStructures
 open DataStructures
 
@@ -14,66 +14,58 @@ let dataPath = sprintf @"%s/iris-full.txt" baseDatasetsLocation
 let baseModelsPath = @"../../../../MLModels"
 let modelPath = sprintf @"%s/IrisModel.zip" baseModelsPath
 
-let dataLoader (mlContext : MLContext) =
-    mlContext.Data.TextReader(
-        TextLoader.Arguments(
-            Separator = "\t",
-            HasHeader = true,
-            Column = 
-                [|
-                    TextLoader.Column("Label", Nullable DataKind.R4, 0)
-                    TextLoader.Column("SepalLength", Nullable DataKind.R4, 1)
-                    TextLoader.Column("SepalWidth", Nullable DataKind.R4, 2)
-                    TextLoader.Column("PetalLength", Nullable DataKind.R4, 3)
-                    TextLoader.Column("PetalWidth", Nullable DataKind.R4, 4)
-                |]
-        )
-    )
-
-let read (dataPath : string) (dataLoader : TextLoader) =
-    dataLoader.Read dataPath
-
 
 [<EntryPoint>]
 let main argv =
 
-    let mlContext = MLContext(seed = Nullable 1)
+    //Create the MLContext to share across components for deterministic results
+    let mlContext = MLContext(seed = Nullable 1)    //Seed set to any number so you have a deterministic environment
 
-    //STEP 1: Common data loading
+    // STEP 1: Common data loading configuration
     let fullData = 
-        dataLoader mlContext
-        |> read dataPath
-
-    let struct (trainingDataView, testingDataView) = mlContext.Clustering.TrainTestSplit(fullData, testFraction = 0.2)
+        mlContext.Data.LoadFromTextFile(dataPath,
+            hasHeader = true,
+            separatorChar = '\t',
+            columns =
+                [|
+                    TextLoader.Column("Label", DataKind.Single, 0)
+                    TextLoader.Column("SepalLength", DataKind.Single, 1)
+                    TextLoader.Column("SepalWidth", DataKind.Single, 2)
+                    TextLoader.Column("PetalLength", DataKind.Single, 3)
+                    TextLoader.Column("PetalWidth", DataKind.Single, 4)
+                |]
+        )
+    
+    //Split dataset in two parts: TrainingDataset (80%) and TestDataset (20%)
+    let trainingDataView, testingDataView = 
+        let split = mlContext.Data.TrainTestSplit(fullData, testFraction = 0.2)
+        split.TrainSet, split.TestSet
 
     //STEP 2: Process data transformations in pipeline
-    let dataProcessPipeline =
-        mlContext.Transforms.Concatenate("Features", "SepalLength", "SepalWidth", "PetalLength", "PetalWidth")
-
+    let dataProcessPipeline = 
+        mlContext.Transforms.Concatenate("Features", "SepalLength", "SepalWidth", "PetalLength", "PetalWidth") 
+        |> Common.ConsoleHelper.downcastPipeline
+        
     // (Optional) Peek data in training DataView after applying the ProcessPipeline's transformations  
     Common.ConsoleHelper.peekDataViewInConsole<IrisData> mlContext trainingDataView dataProcessPipeline 10 |> ignore
     Common.ConsoleHelper.peekVectorColumnDataInConsole mlContext "Features" trainingDataView dataProcessPipeline 10 |> ignore
 
-    let trainer = mlContext.Clustering.Trainers.KMeans(features = "Features", clustersCount = 3)
-
-    // STEP 3: Create and train the model                
-    let modelBuilder = 
-        Common.ModelBuilder.create mlContext dataProcessPipeline
-        |> Common.ModelBuilder.addTrainer trainer
-
-    let trainedModel = 
-        modelBuilder
-        |> Common.ModelBuilder.train trainingDataView
+    // STEP 3: Create and train the model     
+    let trainer = mlContext.Clustering.Trainers.KMeans(featureColumnName = "Features", numberOfClusters = 3)
+    let trainingPipeline = dataProcessPipeline.Append(trainer)
+    let trainedModel = trainingPipeline.Fit(trainingDataView)
 
     // STEP4: Evaluate accuracy of the model
-    let metrics = 
-        (trainedModel, modelBuilder)
-        |> Common.ModelBuilder.evaluateClusteringModel testingDataView
+    let (predictions : IDataView) = trainedModel.Transform(testingDataView)
+    let metrics = mlContext.Clustering.Evaluate(predictions, scoreColumnName = "Score", featureColumnName = "Features")
+
     Common.ConsoleHelper.printClusteringMetrics (trainer.ToString()) metrics
 
+
     // STEP5: Save/persist the model as a .ZIP file
-    (trainedModel, modelBuilder)
-    |> Common.ModelBuilder.saveModelAsFile modelPath
+    use fs = new FileStream(modelPath, FileMode.Create, FileAccess.Write, FileShare.Write)
+    mlContext.Model.Save(trainedModel, trainingDataView.Schema, fs)
+    fs.Close()
 
     printfn "=============== End of training process ==============="
 
@@ -88,14 +80,18 @@ let main argv =
             PetalWidth = 5.1f
         }
 
-    //Create the clusters: Create data files and plot a chart
-    let prediction = 
-        Common.ModelScorer.create mlContext
-        |> Common.ModelScorer.loadModelFromZipFile modelPath
-        |> Common.ModelScorer.predictSingle sampleIrisData
+    use stream = new FileStream(modelPath, FileMode.Open, FileAccess.Read, FileShare.Read)
+    let model,inputSchema = mlContext.Model.Load(stream)
+    // Create prediction engine related to the loaded trained model
+    let predEngine = mlContext.Model.CreatePredictionEngine<IrisData, IrisPrediction>(model)
 
-    printfn "Cluster assigned for setosa flowers: %d" prediction.SelectedClusterId
+    //Score
+    let resultprediction = predEngine.Predict(sampleIrisData)
 
-    Common.ConsoleHelper.consolePressAnyKey ()
+    printfn "Cluster assigned for setosa flowers: %d" resultprediction.SelectedClusterId
+
+    printfn "=============== End of process, hit any key to finish ==============="
+
+    Console.ReadKey() |> ignore
 
     0 // return an integer exit code
