@@ -3,10 +3,9 @@
 open System
 open System.IO
 open Microsoft.ML
-open Microsoft.ML.Runtime.Data
-open MulticlassClassification_Iris.DataStructures
-open Common
 open MulticlassClassification_Iris
+open MulticlassClassification_Iris.DataStructures
+open Microsoft.ML.Data
 
 let appPath = Path.GetDirectoryName(Environment.GetCommandLineArgs().[0])
 
@@ -18,94 +17,72 @@ let baseModelsPath = @"../../../../MLModels"
 let modelPath = sprintf @"%s/IrisClassificationModel.zip" baseModelsPath
 
 
-let dataLoader (mlContext : MLContext) =
-    mlContext.Data.TextReader(
-        TextLoader.Arguments(
-            Separator = "tab",
-            HasHeader = true,
-            Column = 
-                [|
-                    TextLoader.Column("Label", Nullable DataKind.R4, 0)
-                    TextLoader.Column("SepalLength", Nullable DataKind.R4, 1)
-                    TextLoader.Column("SepalWidth", Nullable DataKind.R4, 2)
-                    TextLoader.Column("PetalLength", Nullable DataKind.R4, 3)
-                    TextLoader.Column("PetalWidth", Nullable DataKind.R4, 4)
-                |]
-        )
-    )
-
-let read (dataPath : string) (dataLoader : TextLoader) =
-    dataLoader.Read dataPath
-
 let buildTrainEvaluateAndSaveModel (mlContext : MLContext) =
     
     // STEP 1: Common data loading configuration
-    let trainingDataView = 
-        dataLoader mlContext
-        |> read trainDataPath
-
-    let testDataView = 
-        dataLoader mlContext
-        |> read testDataPath
+    let trainingDataView = mlContext.Data.LoadFromTextFile<IrisData>(trainDataPath, hasHeader = true)
+    let testDataView = mlContext.Data.LoadFromTextFile<IrisData>(testDataPath, hasHeader = true)
 
     // STEP 2: Common data process configuration with pipeline data transformations
-    let dataProcessPipeline =
-        mlContext.Transforms.Concatenate("Features", [| "SepalLength"; "SepalWidth"; "PetalLength"; "PetalWidth"|])
+    let dataProcessPipeline = 
+        EstimatorChain()
+            .Append(mlContext.Transforms.Conversion.MapValueToKey("LabelKey","Label"))
+            .Append(mlContext.Transforms.Concatenate("Features", "SepalLength",
+                                                     "SepalWidth",
+                                                     "PetalLength",
+                                                     "PetalWidth"))
+            .AppendCacheCheckpoint(mlContext)
 
-    // (OPTIONAL) Peek data (such as 5 records) in training DataView after applying the ProcessPipeline's transformations into "Features" 
-    Common.ConsoleHelper.peekDataViewInConsole<IrisData> mlContext trainingDataView dataProcessPipeline 5 |> ignore
-    Common.ConsoleHelper.peekVectorColumnDataInConsole mlContext "Features" trainingDataView dataProcessPipeline 5 |> ignore
-
-    // STEP 3: Set the training algorithm, then create and config the modelBuilder
-    let trainer = mlContext.MulticlassClassification.Trainers.StochasticDualCoordinateAscent(label = "Label", features = "Features")
-    let modelBuilder = 
-        Common.ModelBuilder.create mlContext dataProcessPipeline
-        |> Common.ModelBuilder.addTrainer trainer
+    // STEP 3: Set the training algorithm, then append the trainer to the pipeline  
+    let trainer = mlContext.MulticlassClassification.Trainers.SdcaMaximumEntropy(labelColumnName = "LabelKey", featureColumnName = "Features")
+    let trainingPipeline = dataProcessPipeline.Append(trainer)
 
     // STEP 4: Train the model fitting to the DataSet
     printfn "=============== Training the model ==============="
-    let trainedModel = 
-        modelBuilder
-        |> Common.ModelBuilder.train trainingDataView
+    let trainedModel = trainingPipeline.Fit(trainingDataView)
 
     // STEP 5: Evaluate the model and show accuracy stats
     printfn "===== Evaluating Model's accuracy with Test data ====="
-    let metrics = 
-        (trainedModel, modelBuilder)
-        |> Common.ModelBuilder.evaluateMultiClassClassificationModel testDataView "Label" "Score"
+    let predictions = trainedModel.Transform(testDataView)
+    let metrics = mlContext.MulticlassClassification.Evaluate(predictions, "Label", "Score")
 
     Common.ConsoleHelper.printMultiClassClassificationMetrics (trainer.ToString()) metrics
 
     // STEP 6: Save/persist the trained model to a .ZIP file
-    printfn "=============== Saving the model to a file ==============="
-    (trainedModel, modelBuilder)
-    |> Common.ModelBuilder.saveModelAsFile modelPath
+    use fs = new FileStream(modelPath, FileMode.Create, FileAccess.Write, FileShare.Write)
+    mlContext.Model.Save(trainedModel, trainingDataView.Schema, fs);
 
+    printfn "The model is saved to %s" modelPath
 
 
 let testSomePredictions (mlContext : MLContext) =
-
     //Test Classification Predictions with some hard-coded samples 
-    let modelScorer = 
-        Common.ModelScorer.create mlContext
-        |> Common.ModelScorer.loadModelFromZipFile modelPath
-        
-    let prediction = modelScorer |> Common.ModelScorer.predictSingle DataStructures.TestIrisData.Iris1
-    printfn "Actual: setosa.     Predicted probability: setosa:      %.4f" prediction.Score.[0]
-    printfn "                                           versicolor:  %.4f" prediction.Score.[1]
-    printfn "                                           virginica:   %.4f" prediction.Score.[2]
+    use stream = new FileStream(modelPath, FileMode.Open, FileAccess.Read, FileShare.Read)
+    let trainedModel, inputSchema = mlContext.Model.Load(stream);
+
+    // Create prediction engine related to the loaded trained model
+    let predEngine = mlContext.Model.CreatePredictionEngine<IrisData, IrisPrediction>(trainedModel)
+
+    //Score sample 1
+    let resultprediction1 = predEngine.Predict(DataStructures.SampleIrisData.Iris1)
+
+    printfn "Actual: setosa.     Predicted probability: setosa:      %.4f" resultprediction1.Score.[0]
+    printfn "                                           versicolor:  %.4f" resultprediction1.Score.[1]
+    printfn "                                           virginica:   %.4f" resultprediction1.Score.[2]
     printfn ""
 
-    let prediction = modelScorer |> Common.ModelScorer.predictSingle DataStructures.TestIrisData.Iris2
-    printfn "Actual: virginica.  Predicted probability: setosa:      %.4f" prediction.Score.[0]
-    printfn "                                           versicolor:  %.4f" prediction.Score.[1]
-    printfn "                                           virginica:   %.4f" prediction.Score.[2]
+    //Score sample 2
+    let resultprediction2 = predEngine.Predict(DataStructures.SampleIrisData.Iris2);
+    printfn "Actual: virginica.  Predicted probability: setosa:      %.4f" resultprediction2.Score.[0]
+    printfn "                                           versicolor:  %.4f" resultprediction2.Score.[1]
+    printfn "                                           virginica:   %.4f" resultprediction2.Score.[2]
     printfn ""
 
-    let prediction = modelScorer |> Common.ModelScorer.predictSingle DataStructures.TestIrisData.Iris3
-    printfn "Actual: versicolor. Predicted probability: setosa:      %.4f" prediction.Score.[0]
-    printfn "                                           versicolor:  %.4f" prediction.Score.[1]
-    printfn "                                           virginica:   %.4f" prediction.Score.[2]
+    //Score sample 3
+    let resultprediction2 = predEngine.Predict(DataStructures.SampleIrisData.Iris3);
+    printfn "Actual: versicolor. Predicted probability: setosa:      %.4f" resultprediction2.Score.[0]
+    printfn "                                           versicolor:  %.4f" resultprediction2.Score.[1]
+    printfn "                                           virginica:   %.4f" resultprediction2.Score.[2]
     printfn ""
 
 
@@ -123,6 +100,6 @@ let main argv =
     testSomePredictions mlContext
 
     printfn "=============== End of process, hit any key to finish ==============="
-    ConsoleHelper.consolePressAnyKey()
+    Console.ReadKey() |> ignore
 
     0 // return an integer exit code
