@@ -1,9 +1,9 @@
-﻿// Learn more about F# at http://fsharp.org
-
+﻿
 open System
 open Microsoft.ML
-open Microsoft.ML.Core.Data
-open Microsoft.ML.Runtime.Data
+open System.IO
+open DataStructures
+open Microsoft.ML.Data
 
 let modelsLocation = @"../../../../MLModels"
 
@@ -11,33 +11,15 @@ let datasetsLocation = @"../../../../Data"
 let trainingDataLocation = sprintf @"%s/hour_train.csv" datasetsLocation
 let testDataLocation = sprintf @"%s/hour_test.csv" datasetsLocation
 
-let dataLoader (mlContext : MLContext) =
-    mlContext.Data.TextReader(
-        TextLoader.Arguments(
-            Separator = ",",
-            HasHeader = true,
-            Column = 
-                [|
-                    TextLoader.Column("Season", Nullable DataKind.R4, 2)
-                    TextLoader.Column("Year", Nullable DataKind.R4, 3)
-                    TextLoader.Column("Month", Nullable DataKind.R4, 4)
-                    TextLoader.Column("Hour", Nullable DataKind.R4, 5)
-                    TextLoader.Column("Holiday", Nullable DataKind.R4, 6)
-                    TextLoader.Column("Weekday", Nullable DataKind.R4, 7)
-                    TextLoader.Column("WorkingDay", Nullable DataKind.R4, 8)
-                    TextLoader.Column("Weather", Nullable DataKind.R4, 9)
-                    TextLoader.Column("Temperature", Nullable DataKind.R4, 10)
-                    TextLoader.Column("NormalizedTemperature", Nullable DataKind.R4, 11)
-                    TextLoader.Column("Humidity", Nullable DataKind.R4, 12)
-                    TextLoader.Column("Windspeed", Nullable DataKind.R4, 13)
-                    TextLoader.Column("Count", Nullable DataKind.R4, 16)
-                |]
-        )
-    )
+/// Cast ML.NET pipeline object to IEstimator<ITransformer> interface
+let downcastPipeline (pipeline : IEstimator<'a>) =
+    match pipeline with
+    | :? IEstimator<ITransformer> as p -> p
+    | _ -> failwith "The pipeline has to be an instance of IEstimator<ITransformer>."
 
-let read (dataPath : string) (dataLoader : TextLoader) =
-    dataLoader.Read dataPath
-
+let absolutePath relativePath = 
+    let dataRoot = FileInfo(Reflection.Assembly.GetExecutingAssembly().Location)
+    Path.Combine(dataRoot.Directory.FullName, relativePath)
 
 [<EntryPoint>]
 let main argv =
@@ -47,37 +29,33 @@ let main argv =
     let mlContext = MLContext(seed = Nullable 0)
 
     // 1. Common data loading configuration
-    let trainingDataView = 
-        dataLoader mlContext
-        |> read trainingDataLocation
-
-    let testDataView = 
-        dataLoader mlContext
-        |> read testDataLocation
+    let trainingDataView = mlContext.Data.LoadFromTextFile<DemandObservation>(trainingDataLocation, hasHeader = true, separatorChar = ',')
+    let testDataView = mlContext.Data.LoadFromTextFile<DemandObservation>(testDataLocation, hasHeader = true, separatorChar = ',')
 
 
     // 2: Common data process configuration with pipeline data transformations
     let dataProcessPipeline =
-        mlContext.Transforms.CopyColumns("Count", "Label")
-        |> Common.ModelBuilder.append(
-            mlContext.Transforms.Concatenate("Features", "Season", "Year", "Month",
-                                            "Hour", "Holiday", "Weekday",
-                                            "Weather", "Temperature", "NormalizedTemperature",
-                                            "Humidity", "Windspeed"))
-        |> Common.ConsoleHelper.downcastPipeline
+        EstimatorChain()
+            .Append(mlContext.Transforms.Concatenate("Features", "Season", "Year", "Month",
+                                                     "Hour", "Holiday", "Weekday", "WorkingDay",
+                                                     "Weather", "Temperature", "NormalizedTemperature",
+                                                     "Humidity", "Windspeed"))
+            .AppendCacheCheckpoint(mlContext)
+            // Use in-memory cache for small/medium datasets to lower training time. 
+            // Do NOT use it (remove .AppendCacheCheckpoint()) when handling very large datasets.
+        |> downcastPipeline
 
-    // (OPTIONAL) Peek data (such as 2 records) in training DataView after applying the ProcessPipeline's transformations into "Features" 
+    // (Optional) Peek data in training DataView after applying the ProcessPipeline's transformations  
     Common.ConsoleHelper.peekDataViewInConsole<DataStructures.DemandObservation> mlContext trainingDataView dataProcessPipeline 10 |> ignore
     Common.ConsoleHelper.peekVectorColumnDataInConsole mlContext "Features" trainingDataView dataProcessPipeline 10 |> ignore
 
     // Definition of regression trainers/algorithms to use
-    //var regressionLearners = new (string name, IEstimator<ITransformer> value)[]
     let regressionLearners : (string * IEstimator<ITransformer>) array =
         [|
-            "FastTree", mlContext.Regression.Trainers.FastTree() |> Common.ConsoleHelper.downcastPipeline
-            "Poisson", mlContext.Regression.Trainers.PoissonRegression() |> Common.ConsoleHelper.downcastPipeline
-            "SDCA", mlContext.Regression.Trainers.StochasticDualCoordinateAscent() |> Common.ConsoleHelper.downcastPipeline
-            "FastTreeTweedie", mlContext.Regression.Trainers.FastTreeTweedie() |> Common.ConsoleHelper.downcastPipeline
+            "FastTree", mlContext.Regression.Trainers.FastTree() |> downcastPipeline
+            "Poisson", mlContext.Regression.Trainers.LbfgsPoissonRegression() |> downcastPipeline
+            "SDCA", mlContext.Regression.Trainers.Sdca() |> downcastPipeline
+            "FastTreeTweedie", mlContext.Regression.Trainers.FastTreeTweedie() |> downcastPipeline
             //Other possible learners that could be included
             //...FastForestRegressor...
             //...GeneralizedAdditiveModelRegressor...
@@ -87,39 +65,40 @@ let main argv =
     // 3. Phase for Training, Evaluation and model file persistence
     // Per each regression trainer: Train, Evaluate, and Save a different model
     for (learnerName, trainer) in regressionLearners do
-        printfn "================== Training model =================="
-        let modelBuilder = 
-            Common.ModelBuilder.create mlContext dataProcessPipeline
-            |> Common.ModelBuilder.addTrainer trainer
-
-        let trainedModel = 
-            modelBuilder
-            |> Common.ModelBuilder.train trainingDataView
-            
+        printfn "=============== Training the current model ==============="
+        let trainingPipeline = dataProcessPipeline.Append(trainer)
+        let trainedModel = trainingPipeline.Fit(trainingDataView)
+        
         printfn "===== Evaluating Model's accuracy with Test data ====="
-        let metrics = 
-            (trainedModel, modelBuilder)
-            |> Common.ModelBuilder.evaluateRegressionModel testDataView "Count" "Score"
+        let predictions = trainedModel.Transform(testDataView)
+        let metrics = mlContext.Regression.Evaluate(predictions, labelColumnName = "Label", scoreColumnName = "Score")
+        Common.ConsoleHelper.printRegressionMetrics (trainer.ToString()) metrics
+        
 
-        Common.ConsoleHelper.printRegressionMetrics learnerName metrics
 
         //Save the model file that can be used by any application
-        (trainedModel, modelBuilder)
-            |> Common.ModelBuilder.saveModelAsFile (sprintf @"%s/%sModel.zip" modelsLocation learnerName)
+        let modelPath = sprintf "%s/%sModel.zip" modelsLocation learnerName
+        use fs = new FileStream(modelPath, FileMode.Create, FileAccess.Write, FileShare.Write)
+        mlContext.Model.Save(trainedModel, trainingDataView.Schema, fs)
+
+        printfn "The model is saved to %s" (absolutePath modelPath)
  
     // 4. Try/test Predictions with the created models
     // The following test predictions could be implemented/deployed in a different application (production apps)
     // that's why it is seggregated from the previous loop
     // For each trained model, test 10 predictions           
     for (learnerName, _) in regressionLearners do
-        //Load current model
-        let modelScorer = 
-            Common.ModelScorer.create mlContext
-            |> Common.ModelScorer.loadModelFromZipFile (sprintf @"%s/%sModel.zip" modelsLocation learnerName)
+        //Load current model from .ZIP file
+        let modelPath = sprintf "%s/%sModel.zip" modelsLocation learnerName
+        use stream = new FileStream(modelPath, FileMode.Open, FileAccess.Read, FileShare.Read)
+        let trainedModel, inputSchema = mlContext.Model.Load stream
 
+        // Create prediction engine related to the loaded trained model
+        let predEngine = mlContext.Model.CreatePredictionEngine<DemandObservation, DemandPrediction>(trainedModel)
         printfn "================== Visualize/test 10 predictions for model %sModel.zip ==================" learnerName
+
         //Visualize 10 tests comparing prediction with actual/observed values from the test dataset
-        ModelScoringTester.visualizeSomePredictions testDataLocation modelScorer 10
+        ModelScoringTester.visualizeSomePredictions testDataLocation predEngine 10
 
     Common.ConsoleHelper.consolePressAnyKey ()
 
