@@ -1,12 +1,12 @@
 ﻿open System
 open System.IO
+open System.Linq
 open System.Diagnostics
 open PLplot
 open TaxiFarePrediction.DataStructures.DataStructures
 open Microsoft.ML
-open Microsoft.ML.Runtime.Data
+open Microsoft.ML.Data
 open Microsoft.ML.Transforms
-open Microsoft.ML.Transforms.Normalizers
 
 let appPath = Path.GetDirectoryName(Environment.GetCommandLineArgs().[0])
 
@@ -17,78 +17,63 @@ let testDataPath = sprintf @"%s/taxi-fare-test.csv" baseDatasetsLocation
 let baseModelsPath = @"../../../../MLModels"
 let modelPath = sprintf @"%s/TaxiFareModel.zip" baseModelsPath
 
-let dataLoader (mlContext : MLContext) =
-    mlContext.Data.TextReader(
-        TextLoader.Arguments(
-            Separator = ",",
-            HasHeader = true,
-            Column = 
-                [|
-                    TextLoader.Column("VendorId", Nullable DataKind.Text, 0)
-                    TextLoader.Column("RateCode", Nullable DataKind.Text, 1)
-                    TextLoader.Column("PassengerCount", Nullable DataKind.R4, 2)
-                    TextLoader.Column("TripTime", Nullable DataKind.R4, 3)
-                    TextLoader.Column("TripDistance", Nullable DataKind.R4, 4)
-                    TextLoader.Column("PaymentType", Nullable DataKind.Text, 5)
-                    TextLoader.Column("FareAmount", Nullable DataKind.R4, 6)
-                |]
-        )
-    )
+let downcastPipeline (x : IEstimator<_>) = 
+    match x with 
+    | :? IEstimator<ITransformer> as y -> y
+    | _ -> failwith "downcastPipeline: expecting a IEstimator<ITransformer>"
 
-let read (dataPath : string) (dataLoader : TextLoader) =
-    dataLoader.Read dataPath
 
 let buildTrainEvaluateAndSaveModel (mlContext : MLContext) =
-    
-    // STEP 1: Common data loading configuration
-    let trainingDataView = 
-        dataLoader mlContext
-        |> read trainDataPath
 
-    let testDataView = 
-        dataLoader mlContext
-        |> read testDataPath
+    // STEP 1: Common data loading configuration
+    let baseTrainingDataView = mlContext.Data.LoadFromTextFile<TaxiTrip>(trainDataPath, hasHeader = true, separatorChar = ',')
+    let testDataView = mlContext.Data.LoadFromTextFile<TaxiTrip>(testDataPath, hasHeader = true, separatorChar = ',')
+
+    //Sample code of removing extreme data like "outliers" for FareAmounts higher than $150 and lower than $1 which can be error-data 
+    //let cnt = baseTrainingDataView.GetColumn<decimal>(mlContext, "FareAmount").Count()
+    let trainingDataView = mlContext.Data.FilterRowsByColumn(baseTrainingDataView, "FareAmount", lowerBound = 1., upperBound = 150.)
+    //let cnt2 = trainingDataView.GetColumn<float>(mlContext, "FareAmount").Count()
 
     // STEP 2: Common data process configuration with pipeline data transformations
     let dataProcessPipeline =
-        CopyColumnsEstimator(mlContext, "FareAmount", "Label")
-        |> Common.ModelBuilder.append(mlContext.Transforms.Categorical.OneHotEncoding("VendorId", "VendorIdEncoded"))
-        |> Common.ModelBuilder.append(mlContext.Transforms.Categorical.OneHotEncoding("RateCode", "RateCodeEncoded"))
-        |> Common.ModelBuilder.append(mlContext.Transforms.Categorical.OneHotEncoding("PaymentType", "PaymentTypeEncoded"))
-        |> Common.ModelBuilder.append(mlContext.Transforms.Normalize(inputName = "PassengerCount", mode = NormalizingEstimator.NormalizerMode.MeanVariance))
-        |> Common.ModelBuilder.append(mlContext.Transforms.Normalize(inputName = "TripTime", mode = NormalizingEstimator.NormalizerMode.MeanVariance))
-        |> Common.ModelBuilder.append(mlContext.Transforms.Normalize(inputName = "TripDistance", mode = NormalizingEstimator.NormalizerMode.MeanVariance))
-        |> Common.ModelBuilder.append(mlContext.Transforms.Concatenate("Features", "VendorIdEncoded", "RateCodeEncoded", "PaymentTypeEncoded", "PassengerCount", "TripTime", "TripDistance"))
-        |> Common.ConsoleHelper.downcastPipeline
+        EstimatorChain()
+            .Append(mlContext.Transforms.CopyColumns("Label", "FareAmount"))
+            .Append(mlContext.Transforms.Categorical.OneHotEncoding("VendorIdEncoded", "VendorId"))
+            .Append(mlContext.Transforms.Categorical.OneHotEncoding("RateCodeEncoded", "RateCode"))
+            .Append(mlContext.Transforms.Categorical.OneHotEncoding("PaymentTypeEncoded", "PaymentType"))
+            .Append(mlContext.Transforms.NormalizeMeanVariance("PassengerCount", "PassengerCount"))
+            .Append(mlContext.Transforms.NormalizeMeanVariance("TripTime", "TripTime"))
+            .Append(mlContext.Transforms.NormalizeMeanVariance("TripDistance", "TripDistance"))
+            .Append(mlContext.Transforms.Concatenate("Features", "VendorIdEncoded", "RateCodeEncoded", "PaymentTypeEncoded", "PassengerCount", "TripTime", "TripDistance"))
+            .AppendCacheCheckpoint(mlContext)
+            |> downcastPipeline
 
     // (OPTIONAL) Peek data (such as 5 records) in training DataView after applying the ProcessPipeline's transformations into "Features" 
     Common.ConsoleHelper.peekDataViewInConsole<TaxiTrip> mlContext trainingDataView dataProcessPipeline 5 |> ignore
     Common.ConsoleHelper.peekVectorColumnDataInConsole mlContext "Features" trainingDataView dataProcessPipeline 5 |> ignore
 
-    // STEP 3: Set the training algorithm, then create and config the modelBuilder
-    let trainer = mlContext.Regression.Trainers.StochasticDualCoordinateAscent(label = "Label", features = "Features")
-    let modelBuilder = 
-        Common.ModelBuilder.create mlContext dataProcessPipeline
-        |> Common.ModelBuilder.addTrainer trainer
+    // STEP 3: Set the training algorithm, then create and config the modelBuilder - Selected Trainer (SDCA Regression algorithm)                            
+    let trainer = mlContext.Regression.Trainers.Sdca(labelColumnName = "Label", featureColumnName = "Features")
+
+    let modelBuilder = dataProcessPipeline.Append trainer
 
     // STEP 4: Train the model fitting to the DataSet
+    //The pipeline is trained on the dataset that has been loaded and transformed.
     printfn "=============== Training the model ==============="
-    let trainedModel = 
-        modelBuilder
-        |> Common.ModelBuilder.train trainingDataView
+    let trainedModel = modelBuilder.Fit trainingDataView
 
     // STEP 5: Evaluate the model and show accuracy stats
     printfn "===== Evaluating Model's accuracy with Test data ====="
     let metrics = 
-        (trainedModel, modelBuilder)
-        |> Common.ModelBuilder.evaluateRegressionModel testDataView "Label" "Score"
+        let predictions = trainedModel.Transform testDataView
+        mlContext.Regression.Evaluate(predictions, "Label", "Score")
 
     Common.ConsoleHelper.printRegressionMetrics (trainer.ToString()) metrics
 
     // STEP 6: Save/persist the trained model to a .ZIP file
     printfn "=============== Saving the model to a file ==============="
-    (trainedModel, modelBuilder)
-    |> Common.ModelBuilder.saveModelAsFile modelPath
+    use fs = new FileStream(modelPath, FileMode.Create, FileAccess.Write, FileShare.Write)
+    mlContext.Model.Save(trainedModel, trainingDataView.Schema, fs)
 
 
 let testSinglePrediction (mlContext : MLContext) =
@@ -107,9 +92,11 @@ let testSinglePrediction (mlContext : MLContext) =
         };
 
     let resultprediction = 
-        Common.ModelScorer.create mlContext
-        |> Common.ModelScorer.loadModelFromZipFile modelPath
-        |> Common.ModelScorer.predictSingle taxiTripSample
+        let model, inputSchema = 
+            use s = File.OpenRead(modelPath)
+            mlContext.Model.Load(s)
+        let predictionFunction = mlContext.Model.CreatePredictionEngine(model)
+        predictionFunction.Predict taxiTripSample
 
     printfn "=============== Single Prediction  ==============="
     printfn "Predicted fare: %.4f, actual fare: 15.5" resultprediction.FareAmount
@@ -134,9 +121,11 @@ let plotRegressionChart (mlContext : MLContext) testDataSetPath numberOfRecordsT
             )
         |> Array.take numMaxRecords
     
-    let modelScorer = 
-        Common.ModelScorer.create mlContext
-        |> Common.ModelScorer.loadModelFromZipFile modelPath
+    let modelScorer, inputeSchema = 
+        use s = File.OpenRead(modelPath)
+        mlContext.Model.Load(s)
+
+    let predFunction = mlContext.Model.CreatePredictionEngine(modelScorer)
 
     use pl = new PLStream()
     // use SVG backend and write to SineWaves.svg in current directory
@@ -192,7 +181,7 @@ let plotRegressionChart (mlContext : MLContext) testDataSetPath numberOfRecordsT
     for i in 0 .. testData.Length - 1 do
 
         //Make Prediction
-        let farePrediction = modelScorer |> Common.ModelScorer.predictSingle testData.[i]
+        let farePrediction = predFunction.Predict testData.[i]
   
         let x = [| float testData.[i].FareAmount |]
         let y = [| float farePrediction.FareAmount |]
